+++ app/src/main/java/com/survivalcoding/todolist/ui/main/TodoAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.survivalcoding.todolist.databinding.TodoListItemBinding
-import com.survivalcoding.todolist.model.TodoItem
+import com.survivalcoding.todolist.data.model.TodoItem
 
 class TodoAdapter(
     private val checkChanged: (Long) -> Unit
