+++ app/src/main/java/com/survivalcoding/todolist/ui/write/SimpleTodoWriteActivity.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.survivalcoding.todolist.databinding.ActivitySimpleTodoWriteBinding
-import com.survivalcoding.todolist.model.TodoItem
+import com.survivalcoding.todolist.data.model.TodoItem
 import com.survivalcoding.todolist.ui.main.MainActivity.Companion.NEW_ID
 
 class SimpleTodoWriteActivity : AppCompatActivity() {
